+++ domain/src/main/java/ai/kaira/domain/assessment/model/AssessmentAnswer.kt
@@ -3,8 +3,8 @@ package ai.kaira.domain.assessment.model
 class AssessmentAnswer(var id: Int,
                        var title: String,
                        var value: Int,
-                       var duration: Long,
-                       var time: Long){
+                       var duration: Double,
+                       var time: Double){
 
     fun clone():AssessmentAnswer{
         return AssessmentAnswer(id, title, value, duration, time)
